requirements.txt
streamlit
SQLAlchemy
pandas
bcrypt
psycopg2-binary
python-dotenv
qrcode[pil]
openpyxl


